- Fixed bug in the login process; - Fixed bug in the raffle price calculator on the item model; - Added menu to the main page for future categories; - Added the raffle price to the item main page; - Fixed bug in the amount provided for paypal payment; - Removed description from the item preview.

+++ Raffle/Models/LogicModels.cs
@@ -47,7 +47,7 @@ namespace Raffle.Models
 
         public decimal RafflePrice
         {
-            get { return decimal.Round(Price / TotalRaffleCount, 2); }
+            get { return decimal.Round((decimal)Price / TotalRaffleCount, 2); }
         }
 
         public Item()
